// BookPrioritizerAngular/src/app/models/Book.model.ts
export class Book
{
    name: string;
    dateAdded: Date;
    goodReadsRating: number;
    goodReadsVotes: number;
    amazonRating: number;
    amazonVotes: number;
    bookId: number;
    authorId: number;
    authorFirstName: string;
    authorLastName: string;
    authorFirstLastName: string;
    authorLastFirstName: string;
    statusId: number;
    statusName: string;
    year: number;
    hidden: boolean;
    ratingPercentile: number;
    votersPercentile: number;
    score: number;
    isEditing: boolean;
}